moving functions out of qx_sequential_reports
